requirements.txt
aiohttp==3.4.4
async-timeout==3.0.1
attrs==18.2.0
chardet==3.0.4
docutils==0.14
idna==2.7
idna-ssl==1.1.0
lockfile==0.12.2
multidict==4.4.2
python-daemon==2.2.0
setproctitle==1.1.10
yarl==1.2.6
